minor fix at some TODO comments, redundant codes
