+++ src/components/form-factory/form-factory.tsx
@@ -17,7 +17,7 @@ import * as Yup from 'yup';
 // base field
 import { FormBaseFieldMeta } from "./form-base-field/form-base-field-meta";
 import { ISingleRestApiResponse } from "../../utils/rest-api";
-import { JsonResponseType } from "../../state-management/rest-api-redux-factory";
+import { JsonResponseType } from "../../state-management/types/factory-types";
 
 
 export enum ActionButtonType {
