+++ src/pages/user-com-app-page/user-com-app-page.tsx
@@ -11,7 +11,7 @@ import { CrudType, RequestStatus } from "../../utils/rest-api";
 import {
     IObjectStore,
     IObjectAction
-} from "../../state-management/rest-api-redux-factory";
+} from "../../state-management/types/factory-types";
 import { Company, labelTypesMapToCompanyGroupTypes, companyGroupTypes } from "../../data-model/company/company";
 import { Application } from "../../data-model/application/application";
 import { ApplicationStatus } from "../../data-model/application-status/application-status";
